+++ package.json
@@ -8,6 +8,7 @@
     "react-dom": "^17.0.1",
     "react-firebase-hooks": "^3.0.3",
     "react-firestore": "^1.0.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1"
   },
   "devDependencies": {
