+++ src/AddItem.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const AddItem = () => {
+    return (
+        <div>
+            Add Item
+        </div>
+    )
+}
+
+export default AddItem
